+++ includes/class-cao.php
@@ -7,6 +7,8 @@
 
 namespace RVOLA\WOO\CAO;
 
+use WC_Order;
+
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
@@ -113,6 +115,9 @@ class CAO {
 					$old_date        = apply_filters( 'woo_cao_date_order', $old_date, $gateway, $mode );
 					$old_date_format = date( 'Y-m-d 00:00:00', $old_date );
 
+					// Status to cancel
+					$woo_status = $this->woo_status();
+
 					$orders = $wpdb->get_results(
 						$wpdb->prepare(
 							"
@@ -121,7 +126,7 @@ class CAO {
 							INNER JOIN $wpdb->postmeta as meta
 							ON posts.ID = meta.post_id
 							WHERE posts.post_type = 'shop_order'
-							AND posts.post_status = 'wc-on-hold'
+							AND posts.post_status IN ('$woo_status')
 							AND posts.post_date < %s
 							AND meta.meta_key = '_payment_method'
 							AND meta.meta_value = %s
@@ -153,7 +158,7 @@ class CAO {
 	 * @param int $order_id order ID.
 	 */
 	private function cancel_order( $order_id ) {
-		$order = new \WC_Order( $order_id );
+		$order = new WC_Order( $order_id );
 
 		$order->update_status(
 			'cancelled',
@@ -163,6 +168,36 @@ class CAO {
 
 	}
 
+	private function woo_status() {
+		$woo_status            = array();
+		$woo_status_authorized = array(
+			'pending',
+			'on-hold',
+			'processing',
+			'completed',
+			'refunded',
+			'failed'
+		);
+
+		$default_status = apply_filters( 'woo_cao_statustocancel', array( 'on-hold' ) );
+
+		if ( $default_status && is_array( $default_status ) ) {
+			foreach ( $default_status as $status ) {
+				if ( in_array( $status, $woo_status_authorized ) ) {
+					$woo_status[] = sprintf( 'wc-%s', $status );
+				}
+			}
+		}
+
+		if ( empty( $woo_status ) ) {
+			$woo_status[] = 'wc-on-hold';
+		}
+
+		$woo_status = implode( "','", $woo_status );
+
+		return $woo_status;
+	}
+
 	/**
 	 * Will check and store the products of the canceled order.
 	 *
@@ -170,7 +205,7 @@ class CAO {
 	 */
 	private function restock( $order_id ) {
 
-		$order = new \WC_Order( $order_id );
+		$order = new WC_Order( $order_id );
 
 		$line_items = $order->get_items();